+++ src/app/components/email-register-confirm/email-register-confirm.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { EmailRegisterConfirmComponent } from './email-register-confirm.component';
+
+@NgModule({
+  imports: [CommonModule],
+  declarations: [EmailRegisterConfirmComponent],
+  exports: [EmailRegisterConfirmComponent]
+})
+export class EmailRegisterConfirmModule {}