+++ src/app/components/email-verification/email-verification.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { MatButtonModule, MatFormFieldModule, MatInputModule } from '@angular/material';
+import { EmailVerificationComponent } from './email-verification.component';
+import { SpinnerModule } from '../spinner/spinner.module';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FontAwesomeModule,
+    MatButtonModule,
+    MatFormFieldModule,
+    MatInputModule,
+    ReactiveFormsModule,
+    RouterModule,
+    SpinnerModule
+  ],
+  declarations: [EmailVerificationComponent],
+  exports: [EmailVerificationComponent]
+})
+export class EmailVerificationModule {}