+++ src/app/config/app.info.ts
@@ -1,6 +1,6 @@
 export const AppInfo = {
   TITLE: 'Bodgam',
-  VERSION: '0.1.1',
+  VERSION: '0.1.1.2',
   AUTHOR: { name: 'Mathias Mille', url: 'https://mathiasmille.fr' },
   STARTYEAR: '2018'
 };