+++ src/app/helpers/error.interceptor.ts
@@ -26,9 +26,9 @@ export class ErrorInterceptor implements HttpInterceptor {
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     return next.handle(request).pipe(
       catchError((err) => {
-        if (err.status === 404) {
+        /* if (err.status === 404) {
           this.router.navigate([err.status]);
-        }
+        } */
 
         const error = err.error.message || err.statusText;
         return throwError(error);