+++ src/app/models/email.model.ts
@@ -1,0 +1,4 @@
+export class Email {
+  subject: string;
+  content: string;
+}