+++ src/app/models/user.model.ts
@@ -1,7 +1,10 @@
+import { Email } from './email.model';
+
 export class User {
   id?: number;
   name: string;
   email: string;
   password?: string;
   token?: string;
+  verificationEmail?: Email;
 }