+++ src/app/services/index.ts
@@ -1,6 +1,7 @@
 export * from './alert.service';
 export * from './board-game-geek.service';
 export * from './country.service';
+export * from './email.service';
 export * from './event.service';
 export * from './event-subscription.service';
 export * from './geocoding.service';